Refactor Codex Connector churn recovery helpers
